feature/BRUI-430_Snake_in_BRUI_Ladebildschirm_view3_view4_einbauen - bundled with esbuild
