+++ package.json
@@ -2,11 +2,13 @@
   "name": "dbi-snake",
   "version": "1.1.0",
   "description": "An open source \"Snake\"-game implementation in pure JavaScript",
-  "main": "./modules/SnakeGame.js",
+  "main": "./lib/main.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "npm run dev",
-    "babel": "babel src -d lib",
+    "bundle": "esbuild src/SnakeGame.js --bundle --outfile=bundle/main.js --loader:.html=text",
+    "babel": "babel bundle -d lib",
+    "package": "npm run bundle babel",
     "dev": "http-server ./ -a localhost -p 1234 -o /index.html"
   },
   "repository": {
@@ -28,6 +30,7 @@
     "@babel/cli": "^7.17.10",
     "@babel/core": "^7.18.5",
     "@babel/preset-env": "^7.18.2",
+    "esbuild": "^0.14.49",
     "http-server": "^14.1.1"
   }
 }
